+++ CHANGELOG.rst
@@ -5,6 +5,10 @@ Current
 -------
 
 - Switch to pytest
+- ``ImageField`` optimization/compression.
+  Resized images are now compressed.
+  Default image can also be optimized on upload with ``FS_IMAGES_OPTIMIZE = True``
+  or by specifying `optimize=True` as field parameter.
 
 0.2.1 (2017-01-17)
 ------------------